added class validation, class TabItemValidation and remaining command UpdateSave and AddSave

+++ TournamentOrganizer/Commands/TeamCommands/DeleteTeamCommand.cs
@@ -1,6 +1,8 @@
-﻿using TournamentOrganizer.BusinessLayer.Service;
+﻿using System.Windows;
+using TournamentOrganizer.BusinessLayer.Service;
 using TournamentOrganizer.BusinessLayer.Service.TeamService;
 using TournamentOrganizer.UI.Command;
+using TournamentOrganizer.UI.Validation.TabItemTeamValidation;
 using TournamentOrganizer.UI.VeiwModels;
 
 namespace TournamentOrganizer.UI.Commands.TeamCommands
@@ -10,14 +12,24 @@ namespace TournamentOrganizer.UI.Commands.TeamCommands
         private readonly TabItemTeamViewModel _viewModel;
 
         private readonly ITeamService _teamService;
+        private TabItemTeamValidation _tabItemTeamValidation;
+
         public DeleteTeamCommand(TabItemTeamViewModel viewModel, ITeamService teamService)
         {
             _viewModel = viewModel;
             _teamService = teamService;
+            _tabItemTeamValidation = new TabItemTeamValidation(viewModel);
         }
 
         public override void Execute(object parameter)
         {
+            if(_tabItemTeamValidation.CheckIsEmptySelectedTeam() == false)
+            {
+                MessageBox.Show("Выберите команду",
+                                "Ошибка ",
+                                MessageBoxButton.OK);
+                return;
+            }
             _teamService.Delete(_viewModel.SelectedTeam.Id);
             _viewModel.Teams.Remove(_viewModel.SelectedTeam);
         }
